Extend storage protocol to support additional S3 API options
The previous version of the protocol only supported a limited subset
of the S3 API functionality. E.g., that prevented us from uploading
content that wasn't file based. We extend the protocol in a compatible
way by adding new arity functions that can take the additional options
as a map.