Add specs to validate implementations of the protocol
Instead of having the specs defined on the implementation, define them
for the core protocol. There are two benefits to this:

- The specs are not tied to a particular implementation, so we avoid
  implementation specific specs that are hard to comply with by other
  implementations.
- We can reuse this core specs for implementation specific additional
  specs.

This new specs also bring a BREAKING CHANGE to the values returned by
the procol methods. They all now return a map with at least a
mandatory boolean key (:success?). They may additionally return other
keys related to their operation (.e.g., the object we wanted to get
from the object storage, a url for such object, etc). And in case the
:success? key is false, there is always a :error-details key that
holds a map with additional details on the problem the method
encountered while trying to perform its operation.