Implement "copy-object" method
[Re #5]

Added implementation of "copy-object" method as described in the related
issue.

For now, as for the Record definition, we do not support copying objects from
different S3 Buckets, as there is no such use-case for it and we prefer to
keep things simple.

We also allow an optional `opts` map for extra options, not used for now but
expected to be used in the future.

Encryption is just supported as a side-effect of copying an already encrypted
object, but we do not support to change this or any other metadata property
for now.

Added new integration tests as well.