Implement missing protocol function arity
A while back ago the core protocol definition added a second arity for
the list-object operation. This second arity was mainly introduced for
those core protocol implementations where listing the objects
recursively is not the default behaviour (like it is in S3).

We fogot to add the arity in the S3 implementation (where it is
basically a no-op).