Ignore clj-kondo "working files"
In addition to the .cache directory, clj-kondo can create other
working directories and files. E.g., directories for 3rd party
dependencies to store their macro linting definitions.

We want to ignore them all, but instead of playing the "whack-a-mole"
game with individual directories and files, we are going to simply
ignore everything by default, and explicitly add the only files that
we actually want to store in the git repository. Namely the clj-kondo
configuration file, and the .gitignore file we use to ignore
everything else.