Avoid warning in put-object with :endpoint and explicit-object-acl
When doing `put-object` operations on an AWSS3Bucket record with both
an `:endpoint` and an `explicit-object-acl`, Amazonica library
generates a warning about not being able to set the S3 bucket account
owner for the ACL.

It turns out the heuristics used by Amazonica to get the account owner
don't work if using an explicit `:endpoint` that doesn't look like the
canonical AWS S3 endpoints. Which happens when using S3 compatible APIs
from other cloud vendors.

Amazonica allows us to explicitly provide the account owner ourselves,
and skips the heuristics (and the warning). So make sure that we
provide the account owner if needed.